end of improvements

+++ src/main/java/dao/CityDao.java
@@ -5,7 +5,7 @@ import entity.CityBean;
 import java.util.List;
 import java.util.stream.Stream;
 
-public interface  CityDao extends Dao<CityBean, Integer> {
+public interface  CityDao extends Dao<CityBean, Long>{
     Stream<CityBean> findByCountryCode(String code);
     Stream<CityBean> findAll();
     List<String> findCountryCodes();
